Tweak Makefile
Some tweaks for the Makefile that allow for more granular testing.  Also
created some .PHONY targets for building the proxy and the client.  This
should be a good start.  More tightening will be required later but I
didn't want to go overboard because we will likely be able to leverage
netavarks Makefile at some point (when we merge?)

Added `validate` target.  This runs cargo fmt and clippy to ensure code
is clean and valid.  Consider running this before each commit.

Signed-off-by: Brent Baude <bbaude@redhat.com>