Update network config fields and implement mac address
Removed the lease field from the network config because it is not needed. This is because the proxy should be keeping track of when to renew a lease and not netavark. 

Mac address is now a string - instead of a byte array. implemented new methods to validate the mac address.

Changed client request to keep track of these changes