Fully attach caching system to gRPC server and finish all gRPC requests
After approaching a memory safety issue of updating the contents of the cache in a different thread. I realized that it was not the fields of the lease cache that need to be atomically referenced mutexes. Instead the instance of the cache was made atomically referenced counted mutexes - in order to share the entire instance with all threads. More detail on why this was necessary can be found in the server binary documentation.

After the cache instance was established, I was able to include other gRPC connection endpoints (teardown(), remove_lease())

Finally I added domain name and host name to the network configuration which was then recorded in the lease.