Send DHCP release on teardown
When we teardown a dhcp macvlan, it is considered good practice to send
a DHCP release message to the server that provided the lease.

changed the return type of remove_lease to return a copy of the lease so
it can be used to send the release message.

added readme for manual testing

Signed-off-by: Brent Baude <bbaude@redhat.com>