Add testing for identical leases
Merge the add and update lease methods into a upsert method. Add and
update used the same logic so it was easier to combine them. Now the
upsert will log if it added or updated a lease.

In update and add lease, try adding the same lease immediately after
upserting the first. This closes issue #50

Signed-off-by: Jack <jackbaude@gmail.com>